Move spectaccular settings to spectacular_settings.py for better separation
