+++ weblate/settings_docker.py
@@ -8,7 +8,7 @@ from logging.handlers import SysLogHandler
 from django.core.exceptions import PermissionDenied
 from django.http import Http404
 
-from weblate.api.spectacular import (
+from weblate.api.spectacular_settings import (
     get_drf_settings,
     get_drf_standardized_errors_settings,
     get_spectacular_settings,
